+++ pyfragment/methods/__init__.py
@@ -3,6 +3,7 @@ from pyfragment.methods.giveaway_premium import giveaway_premium
 from pyfragment.methods.giveaway_stars import giveaway_stars
 from pyfragment.methods.purchase_premium import purchase_premium
 from pyfragment.methods.purchase_stars import purchase_stars
+from pyfragment.methods.recharge_ads import recharge_ads
 from pyfragment.methods.topup_ton import topup_ton
 
 __all__ = [
@@ -11,6 +12,7 @@ __all__ = [
     "giveaway_stars",
     "purchase_premium",
     "purchase_stars",
+    "recharge_ads",
     "terminate_sessions",
     "toggle_login_codes",
     "topup_ton",